refactor(proto): move proto smart function api to v1
