--- crates/jstz_proto/src/api/mod.rs
@@ -1,1 +1,0 @@
-pub mod smart_function;
