+++ crates/jstz_proto/src/runtime/mod.rs
@@ -1,6 +1,6 @@
-mod v1;
+pub mod v1;
 
-pub use v1::{Kv, KvValue, ProtocolApi, ProtocolData};
+pub use v1::{Kv, KvValue, ParsedCode, ProtocolData};
 
 #[cfg(feature = "riscv")]
 pub mod v2;
